understand `<<EOS # => ""`
This sort needs complex parsing if we go strict.  That's what we don't
want to do.  Instead, given we already check for vaidity of the input,
assume this heredoc marker is a part of valid ruby program.  That
assumption ease the situation; we just replace heredoc marker with empty
string when we understand the line.